+++ package.json
@@ -19,6 +19,7 @@
     "react": "^16.8.6",
     "react-app-rewired": "^2.1.1",
     "react-dom": "^16.8.6",
+    "react-full-screen": "^0.2.4",
     "react-loading": "^2.0.3",
     "react-markdown": "^4.0.8",
     "react-monaco-editor": "^0.31.1",
@@ -68,4 +69,3 @@
     "webpack-dev-server": "^3.8.0"
   }
 }
-
